CharacterSelection and LocationScreen additions
Return to previous location functionality added. This is automatically set up when a player enters and leaves a location. It works by storing the current location's init function within a 'hold location' variable that is then redeployed as the 'last location' variable when the next location screen is initialised. from there whenever the player selects the option to go back, the previous location's init function is loaded. 

Now the player can select to view their items, weapons and current statistics from any location screen. It's been baked into the ScreenSystem so that it needn't be repeated in code per location. It works by pulling information from the character class and showing it as a separate screen. The previous screen is then loaded on enter press.